Experimental Vision Recursion Logging
So, the idea is that we log every single AprilTag detection we see (pose, time, tagCount, etc.) along with a reference non-vision position from odometry. Then, using some more or less simply recursion we identify "ideal" standard deviation coefficients for vision. This is purely experimental and needs to be tested extensively. The Recursion will either be in Java or a standalone Python application (I don't know what to do yet).